added experimental scalar product
git-svn-id: https://svn.code.sf.net/p/gaalet/code/trunk@18 d266722a-31f7-4b50-8c79-7894b272e5be

+++ include/cpp0x/configuration_list.h
@@ -4,7 +4,7 @@
 namespace gaalet
 {
 
-typedef unsigned int conf_t;
+typedef unsigned long long int conf_t;
 
 template<conf_t H, typename T>
 struct configuration_list
